Refactoring cherry pick logic out of the dprocedure package and into a new cherry_pick package
